Add Tests for SubtestsResultsView Component (#799)
