+++ src/__tests__/utils/fixtures.ts
@@ -2245,8 +2245,112 @@ const getTestData = () => {
       is_meaningful: true,
     },
   ];
+
+  const subtestsResult = [
+    {
+      base_rev: 'd775409d7c6abb76362a3430e9880ec032ad4679',
+      new_rev: '22f4cf67e8ad76b5ab2a00b97837d1d920b8c2b7',
+      base_app: 'firefox',
+      new_app: 'firefox',
+      framework_id: 1,
+      platform: 'linux1804-64-shippable-qr',
+      suite: 'a11yr',
+      header_name: 'a11yr dhtml.html opt e10s fission stylo webrender',
+      base_repository_name: 'mozilla-central',
+      new_repository_name: 'mozilla-central',
+      is_complete: true,
+      base_measurement_unit: 'ms',
+      new_measurement_unit: 'ms',
+      base_retriggerable_job_ids: [469521067, 469843701],
+      new_retriggerable_job_ids: [469707477, 469843700],
+      base_runs: [964.52, 978.24],
+      new_runs: [977.16, 987.66],
+      base_avg_value: 971.38,
+      new_avg_value: 982.41,
+      base_median_value: 971.38,
+      new_median_value: 982.41,
+      test: 'dhtml.html',
+      option_name: 'opt',
+      extra_options: 'e10s fission stylo webrender',
+      base_stddev: 9.7,
+      new_stddev: 7.42,
+      base_stddev_pct: 1.0,
+      new_stddev_pct: 0.76,
+      confidence: 1.28,
+      confidence_text: 'Low',
+      graphs_link:
+        'https://treeherder.mozilla.org/perfherder/graphs?highlightedRevisions=d775409d7c6a&highlightedRevisions=22f4cf67e8ad&series=mozilla-central%2C5c47b31c38f7214a07fad2e0d41fb901cdc18eae%2C1%2C1&timerange=604800',
+      delta_value: 11.03,
+      delta_percentage: 1.14,
+      magnitude: 5.68,
+      new_is_better: false,
+      lower_is_better: true,
+      is_confident: false,
+      more_runs_are_needed: true,
+      noise_metric: false,
+      is_improvement: false,
+      is_regression: false,
+      is_meaningful: true,
+      base_parent_signature: 4769486,
+      new_parent_signature: 4769486,
+      base_signature_id: 4769487,
+      new_signature_id: 4769487,
+      has_subtests: false,
+    },
+    {
+      base_rev: 'd775409d7c6abb76362a3430e9880ec032ad4679',
+      new_rev: '22f4cf67e8ad76b5ab2a00b97837d1d920b8c2b7',
+      base_app: 'firefox',
+      new_app: 'firefox',
+      framework_id: 1,
+      platform: 'linux1804-64-shippable-qr',
+      suite: 'a11yr',
+      header_name: 'a11yr tablemutation.html opt e10s fission stylo webrender',
+      base_repository_name: 'mozilla-central',
+      new_repository_name: 'mozilla-central',
+      is_complete: true,
+      base_measurement_unit: 'ms',
+      new_measurement_unit: 'ms',
+      base_retriggerable_job_ids: [469521067, 469843701],
+      new_retriggerable_job_ids: [469707477, 469843700],
+      base_runs: [312.44, 320.26],
+      new_runs: [315.78, 323.12],
+      base_avg_value: 316.35,
+      new_avg_value: 319.45,
+      base_median_value: 316.35,
+      new_median_value: 319.45,
+      test: 'tablemutation.html',
+      option_name: 'opt',
+      extra_options: 'e10s fission stylo webrender',
+      base_stddev: 5.53,
+      new_stddev: 5.19,
+      base_stddev_pct: 1.75,
+      new_stddev_pct: 1.62,
+      confidence: 0.58,
+      confidence_text: 'Low',
+      graphs_link:
+        'https://treeherder.mozilla.org/perfherder/graphs?highlightedRevisions=d775409d7c6a&highlightedRevisions=22f4cf67e8ad&series=mozilla-central%2Ccc1b54585c7d1baaadf9d23164e2ad31c9cb07a3%2C1%2C1&timerange=604800',
+      delta_value: 3.1,
+      delta_percentage: 0.98,
+      magnitude: 4.89,
+      new_is_better: false,
+      lower_is_better: true,
+      is_confident: false,
+      more_runs_are_needed: true,
+      noise_metric: false,
+      is_improvement: false,
+      is_regression: false,
+      is_meaningful: true,
+      base_parent_signature: 4769486,
+      new_parent_signature: 4769486,
+      base_signature_id: 4769488,
+      new_signature_id: 4769488,
+      has_subtests: false,
+    },
+  ];
   return {
     testCompareData,
+    subtestsResult,
     testData,
     paginationTestCompareData,
     testCompareDataWithMultipleRuns,
